Replace NovaGL with picaGL for 3DS OpenGL rendering
- Changed video_3ds_gl.c to use picaGL API (pglInit, pglExit, pglSwapBuffers, pglSelectScreen)
- Updated mk/3ds.mk to link -lpicaGL and add include path
- Added picaGL installation step in GitHub Actions workflow
- Fixed include from NovaGL.h to GL/picaGL.h
- Replaced nova_* functions with standard OpenGL + picaGL screen selection
- Top screen: GFX_TOP renders main game view (640x480 scaled to 400x240)
- Bottom screen: GFX_BOTTOM renders zoomed view around cursor with crosshair

+++ src/platform/3ds/storage_3ds.c
@@ -12,7 +12,7 @@
 
 #define SAVE_PATH "sdmc:/3ds/wacki/wacki.sav"
 
-int plat_save_read(void *buf, size_t sz)
+int plat_save_read(void *buf, int sz)
 {
     FILE *f = fopen(SAVE_PATH, "rb");
     if (!f) {
@@ -32,7 +32,7 @@ int plat_save_read(void *buf, size_t sz)
     return 0;
 }
 
-int plat_save_write(const void *buf, size_t sz)
+int plat_save_write(const void *buf, int sz)
 {
     FILE *f = fopen(SAVE_PATH, "wb");
     if (!f) {
